requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
autopep8==2.1.0
beautifulsoup4==4.12.3
black==24.4.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confusable-homoglyphs==3.3.1
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==5.0.6
django-allauth==0.61.1
django-bootstrap-datepicker-plus==5.0.5
django-crispy-forms==2.1
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-pytest==0.2.0
django-registration==3.4
djangorestframework==3.15.1
executing==2.0.1
Faker==24.11.0
faker-vehicle==0.2.0
field==0.2.0
filelock==3.13.4
flake8==7.0.0
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
identify==2.5.36
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
ipython==8.23.0
isort==5.13.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
loguru==0.7.2
lxml==5.2.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mock==5.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
parso==0.8.4
pathspec==0.12.1
phonenumbers==8.13.34
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
Protego==0.3.1
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
Scrapy==2.11.1
selenium==4.21.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
structlog==24.1.0
text-unidecode==1.3
tldextract==5.1.2
tomlkit==0.12.4
traitlets==5.14.2
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
w3lib==2.1.2
waitress==3.0.0
wcwidth==0.2.13
win32-setctime==1.1.0
wsproto==1.2.0
zope.interface==6.3